+++ src/app/loadout-builder/loadout-builder-reducer.ts
